Fix on front addind new features
Since the contract has a way to retrieve the name and balance (i.e.
using contract's public variables) and also a function to set the name
of the contract, the front had to reflect such behavior.

Thus, the front has the contract handler that mimics what you do in the
truffle console.